+++ cmd/darkman/main.go
@@ -2,70 +2,83 @@ package main
 
 import (
 	"fmt"
-	"log"
-	"strings"
+	"os"
 
-	"github.com/integrii/flaggy"
+	"github.com/spf13/cobra"
 	"gitlab.com/WhyNotHugo/darkman"
 	"gitlab.com/WhyNotHugo/darkman/libdarkman"
 )
 
 var Version = "0.0.0-dev"
 
-func NewSubcommand(name, description string) *flaggy.Subcommand {
-	cmd := flaggy.NewSubcommand(name)
-	cmd.Description = description
-
-	flaggy.AttachSubcommand(cmd, 1)
-	return cmd
+// rootCmd represents the base command when called without any subcommands
+var rootCmd = &cobra.Command{
+	Use:     "darkman",
+	Short:   "Query and control darkman from the command line",
+	Version: Version,
 }
 
-func main() {
-	var setVal string
+var setCmd = &cobra.Command{
+	Use:       "set",
+	Short:     "Set the current mode",
+	Args:      cobra.ExactValidArgs(1),
+	ValidArgs: []string{"dark", "light"},
+	RunE: func(cmd *cobra.Command, args []string) error {
+		return libdarkman.SetMode(args[0])
+	},
+}
 
-	get := NewSubcommand("get", "Get and print the current mode")
-	set := NewSubcommand("set", "Set the current mode")
-	toggle := NewSubcommand("toggle", "Toggle the current mode")
-	run := NewSubcommand("run", `Run the darkman service`)
-
-	run.AdditionalHelpPrepend = strings.Join([]string{
-		"",
-		"This command starts the darkman service itself.",
-		"It should only be used by a service manager, by a  session init script or alike.",
-		"",
-		"The service will run in foreground.",
-	}, "\n")
-
-	set.AddPositionalValue(&setVal, "mode", 1, true, "New mode (light or dark)")
-
-	flaggy.SetName("darkman")
-	flaggy.SetVersion(Version)
-	flaggy.SetDescription("Query and control darkman from the command line")
-	flaggy.Parse()
-
-	var err error
-	switch {
-	case get.Used:
+var getCmd = &cobra.Command{
+	Use:   "get",
+	Short: "Get and print the current mode",
+	RunE: func(cmd *cobra.Command, args []string) error {
 		mode, err := libdarkman.GetMode()
 		if err == nil {
 			fmt.Println(mode)
 		}
-	case set.Used:
-		// SetMode validates the mode.
-		err = libdarkman.SetMode(setVal)
-	case toggle.Used:
+		return err
+	},
+}
+
+var toggleCmd = &cobra.Command{
+	Use:   "toggle",
+	Short: "Toggle the current mode and print the new mode",
+	RunE: func(cmd *cobra.Command, args []string) error {
 		mode, err := libdarkman.ToggleMode()
 		if err == nil {
 			fmt.Println(mode)
 		}
+		return err
+	},
+}
 
-	case run.Used:
-		err = darkman.ExecuteService()
-	default:
-		flaggy.ShowHelpAndExit("No command specified")
-	}
+var runCmd = &cobra.Command{
+	Use:   "run",
+	Short: "Run the darkman service",
+	Long: `This command starts the darkman service itself. It should only
+be used by a service manager, by a  session init script or alike.
 
+The service will run in foreground.`,
+	RunE: func(cmd *cobra.Command, args []string) error {
+		// Avoid showing usage if service fails to start.
+		// See: https://github.com/spf13/cobra/issues/340#issuecomment-374617413
+		cmd.SilenceUsage = true
+		return darkman.ExecuteService()
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(setCmd)
+	rootCmd.AddCommand(getCmd)
+	rootCmd.AddCommand(toggleCmd)
+	rootCmd.AddCommand(runCmd)
+}
+
+func main() {
+	err := rootCmd.Execute()
 	if err != nil {
-		log.Fatalf(err.Error())
+		os.Exit(1)
 	}
 }
+
+// TODO: a command to check the configuration file and exit.