+++ CHANGELOG.md
@@ -185,3 +185,5 @@ immediately.
 - Various documentation improvements, including docs for `portals.conf`.
 - Droped hardening rules for the systemd service. This doesn't realistically
   add much security in the end but interferes with several configurations.
+- Switch back from `flaggy` to `cobra`. The latter has fixed the issues that we
+  had in the past, and can generate shell completions.